+++ src/main/scala/com/dataintuitive/viash/cli/RegisteredOpt.scala
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dataintuitive.viash.cli
+package io.viash.cli
 
 case class RegisteredOpt (
   name: String,