+++ src/main/scala/com/dataintuitive/viash/config/package.scala
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dataintuitive.viash
+package io.viash
 
 import io.circe.{Decoder, Encoder, Json}
 import io.circe.generic.extras.Configuration