+++ src/main/scala/com/dataintuitive/viash/config_mods/ConfigModParser.scala
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dataintuitive.viash.config_mods
+package io.viash.config_mods
 
 import io.circe.Json