+++ src/main/scala/com/dataintuitive/viash/functionality/arguments/package.scala
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dataintuitive.viash.functionality
+package io.viash.functionality
 
 import io.circe.{Decoder, Encoder, Json}
 import io.circe.generic.extras.semiauto.{deriveConfiguredDecoder, deriveConfiguredEncoder}
@@ -25,7 +25,7 @@ import java.nio.file.Paths // for .widen
 
 package object arguments {
 
-  import com.dataintuitive.viash.helpers.Circe._
+  import io.viash.helpers.Circe._
 
   implicit val encodeDouble: Encoder[Double] = Encoder.instance {
       value =>