+++ src/main/scala/com/dataintuitive/viash/functionality/resources/NextflowScript.scala
@@ -15,9 +15,9 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dataintuitive.viash.functionality.resources
+package io.viash.functionality.resources
 
-import com.dataintuitive.viash.functionality._
+import io.viash.functionality._
 
 import java.net.URI