+++ src/main/scala/com/dataintuitive/viash/functionality/resources/PlainFile.scala
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dataintuitive.viash.functionality.resources
+package io.viash.functionality.resources
 
 import java.net.URI