+++ src/main/scala/com/dataintuitive/viash/functionality/resources/PythonScript.scala
@@ -15,11 +15,11 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dataintuitive.viash.functionality.resources
+package io.viash.functionality.resources
 
-import com.dataintuitive.viash.functionality._
-import com.dataintuitive.viash.functionality.arguments._
-import com.dataintuitive.viash.wrapper.BashWrapper
+import io.viash.functionality._
+import io.viash.functionality.arguments._
+import io.viash.wrapper.BashWrapper
 
 import java.net.URI