+++ src/main/scala/com/dataintuitive/viash/platforms/NextflowPlatform.scala
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dataintuitive.viash.platforms
+package io.viash.platforms
 
 trait NextflowPlatform extends Platform {
   val variant: String