+++ src/main/scala/com/dataintuitive/viash/platforms/requirements/PythonRequirements.scala
@@ -15,11 +15,11 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dataintuitive.viash.platforms.requirements
+package io.viash.platforms.requirements
 
-import com.dataintuitive.viash.helpers.Circe._
-import com.dataintuitive.viash.helpers.description
-import com.dataintuitive.viash.helpers.example
+import io.viash.helpers.Circe._
+import io.viash.helpers.description
+import io.viash.helpers.example
 
 @description("Specify which Python packages should be available in order to run the component.")
 @example("""setup: