+++ src/main/scala/com/dataintuitive/viash/platforms/requirements/Requirements.scala
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dataintuitive.viash.platforms.requirements
+package io.viash.platforms.requirements
 
 trait Requirements {
   val `type`: String