+++ src/main/scala/com/dataintuitive/viash/platforms/requirements/package.scala
@@ -15,13 +15,13 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dataintuitive.viash.platforms
+package io.viash.platforms
 
 import io.circe.{Decoder, Encoder, Json}
 import io.circe.generic.extras.Configuration
 import io.circe.generic.extras.semiauto.{deriveConfiguredDecoder, deriveConfiguredEncoder}
 import cats.syntax.functor._ // for .widen
-import com.dataintuitive.viash.helpers.Circe._
+import io.viash.helpers.Circe._
 
 package object requirements {
   implicit val encodeRRequirements: Encoder.AsObject[RRequirements] = deriveConfiguredEncoder