+++ src/test/scala/com/dataintuitive/viash/TestHelper.scala
@@ -1,4 +1,4 @@
-package com.dataintuitive.viash
+package io.viash
 
 import java.io.{ByteArrayOutputStream, File, FileInputStream, IOException, UncheckedIOException}
 import java.security.{DigestInputStream, MessageDigest}
@@ -8,11 +8,11 @@ import org.scalatest.Tag
 import java.nio.file.{Files, Path, Paths}
 import scala.reflect.ClassTag
 
-object DockerTest extends Tag("com.dataintuitive.viash.DockerTest")
+object DockerTest extends Tag("io.viash.DockerTest")
 
-object NativeTest extends Tag("com.dataintuitive.viash.NativeTest")
+object NativeTest extends Tag("io.viash.NativeTest")
 
-object NextFlowTest extends Tag("com.dataintuitive.viash.NextFlowTest")
+object NextFlowTest extends Tag("io.viash.NextFlowTest")
 
 object TestHelper {